created звd productsFiltersSlice
