+++ client/src/store/index.js
@@ -1,10 +1,12 @@
 import {combineReducers, configureStore} from "@reduxjs/toolkit";
 import thunk from 'redux-thunk';
 import productsReducer from './productsSlice'
+import productsFiltersReducer from './productsFiltersSlice'
 
 
 const rootReducer = combineReducers({
     products: productsReducer,
+    productsFilters: productsFiltersReducer,
 })
 
 const store = configureStore({
